fix(flagd-evaluation-engine): drop ext-string so custom starts_with/ends_with run
